:truck: [Socialite] Move two utilities into the class
