+++ src/index.ts
@@ -17,8 +17,6 @@ export {
   filterNullishValuesFromObject,
   filterNetworkProperties,
   mergeRegExp,
-  buildUrlFromGroups,
-  fixUrlWithoutScheme,
   getUrlGroups,
   getUrlWithSubstitutions,
 } from './utilities';
