+++ src/utilities/index.ts
@@ -1,9 +1,4 @@
 export {filterNullishValuesFromObject} from './general';
 export {filterNetworkProperties} from './network';
 export {mergeRegExp} from './regexp';
-export {
-  buildUrlFromGroups,
-  fixUrlWithoutScheme,
-  getUrlGroups,
-  getUrlWithSubstitutions,
-} from './url';
+export {getUrlGroups, getUrlWithSubstitutions} from './url';
